third_party/WebKit/LayoutTests/fast/xmlhttprequest/xmlhttprequest-responsetype-sync-request-expected.txt

This tests that the XMLHttpRequest responseType attribute is not modifiable for synchronous HTTP(S) requests.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.responseType = 'document'; threw exception InvalidAccessError: Failed to set the 'responseType' property on 'XMLHttpRequest': the response type can only be changed for asynchronous HTTP requests made from a document..
PASS xhr.responseType = 'document'; threw exception InvalidAccessError: Failed to set the 'responseType' property on 'XMLHttpRequest': the response type can only be changed for asynchronous HTTP requests made from a document..
PASS window.location.protocol is "file:"
xhr.responseType = 'document';
PASS xhr.responseType is "document"
xhr.responseType = 'document';
PASS xhr.responseType is "document"
PASS successfullyParsed is true

TEST COMPLETE
